回転速度があったかも

+++ second_swerve/v1.0/lib/neutral/Neutral.h
@@ -1,8 +1,9 @@
 #pragma once
+#include <stdint.h>
 
 class Neutral {
     public:
-        float setNeutral(float data);
+        float setNeutral(uint8_t data);
     private:
         float ctrlNewtral;
 };
